@@ -14,39 +14,45 @@
     "@agm/core": "^3.0.0-beta.0",
     "@angular-material-extensions/google-maps-autocomplete": "^4.1.1",
     "@angular-material-extensions/password-strength": "^7.0.0",
-    "@angular/animations": "~10.1.2",
-    "@angular/cdk": "^10.2.3",
-    "@angular/common": "~10.1.2",
-    "@angular/compiler": "~10.1.2",
-    "@angular/core": "~10.1.2",
+    "@angular/animations": "^10.2.1",
+    "@angular/cdk": "^10.2.6",
+    "@angular/common": "^10.2.1",
+    "@angular/compiler": "^10.2.1",
+    "@angular/core": "^10.2.1",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "~10.1.2",
-    "@angular/material": "^10.2.3",
-    "@angular/material-moment-adapter": "^10.2.3",
-    "@angular/platform-browser": "~10.1.2",
-    "@angular/platform-browser-dynamic": "~10.1.2",
-    "@angular/router": "~10.1.2",
+    "@angular/forms": "^10.2.1",
+    "@angular/material": "^10.2.6",
+    "@angular/material-moment-adapter": "^10.2.6",
+    "@angular/platform-browser": "^10.2.1",
+    "@angular/platform-browser-dynamic": "^10.2.1",
+    "@angular/router": "^10.2.1",
     "@sweetalert2/ngx-sweetalert2": "^8.1.1",
     "agm-direction": "^0.8.7",
+    "angular-animations": "^0.11.0",
     "angular-oauth2-oidc": "^10.0.3",
     "angular2-promise-buttons": "^4.0.9",
+    "boostrap": "^2.0.0",
+    "bootstrap": "^4.5.3",
     "guid-typescript": "^1.0.9",
-    "jwt-decode": "^3.0.0-beta.2",
-    "moment": "^2.29.0",
+    "jquery": "^3.5.1",
+    "jwt-decode": "^3.0.0",
+    "moment": "^2.29.1",
+    "ng-bootstrap": "^1.6.3",
+    "ngx-lightbox": "^2.2.2",
     "ngx-mask": "^10.0.4",
     "rxjs": "~6.6.0",
-    "sweetalert2": "^10.3.5",
-    "tslib": "^2.0.0",
+    "sweetalert2": "^10.9.0",
+    "tslib": "^2.0.3",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1001.2",
-    "@angular/cli": "~10.1.2",
-    "@angular/compiler-cli": "~10.1.2",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "^0.1001.7",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^6.0.0",
+    "@types/node": "^12.19.3",
+    "codelyzer": "^6.0.1",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~5.0.0",
@@ -57,6 +63,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "typescript": "^4.0.5"
   }
 }
