@@ -11,8 +11,7 @@ import {AgmDirectionModule} from 'agm-direction';
 import {MatGoogleMapsAutocompleteModule} from '@angular-material-extensions/google-maps-autocomplete';
 import {SweetAlert2Module} from '@sweetalert2/ngx-sweetalert2';
 import {MAT_MOMENT_DATE_ADAPTER_OPTIONS, MatMomentDateModule, MomentDateAdapter} from '@angular/material-moment-adapter';
-import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
-import {HttpInterceptor} from '@core/interceptors/httpinterceptor';
+import {HttpClientModule} from '@angular/common/http';
 import {DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE} from '@angular/material/core';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {FooterModule} from './footer/footer.module';
@@ -79,11 +78,6 @@ export function getLocalStorage(): any {
     {
       provide: 'LOCALSTORAGE',
       useFactory: getLocalStorage
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: HttpInterceptor,
-      multi: true
     },
     {
       provide: DateAdapter,
