@@ -1,8 +1,7 @@
 import {Component, HostBinding, Input, OnInit} from '@angular/core';
 import {
-  animate, animateChild,
-  group,
-  query, sequence,
+  animate, group,
+  query,
   style,
   transition,
   trigger,
@@ -13,57 +12,57 @@ import {
   templateUrl: './page-header.component.html',
   styleUrls: ['./page-header.component.scss'],
   animations: [
-    trigger('routerAnimations', [
-      transition('* <=> *', [
-        sequence([
+    trigger(
+      'showHideHeader',
+      [
+        transition(':enter',
           group([
-            /* both variants are important! Do not remove one of them */
-            query('@showPageTitle', animateChild(), {optional: true}),
-            query('@showPageSubTitle', animateChild(), {optional: true})
-          ])
-        ])
-      ]),
-    ]),
-    trigger('showPageTitle', [
-      transition('void => *', [
-        style({
-          opacity: 0,
-          transform: 'translateY(100%)'
-        }),
-        animate('0.4s', style({
-          opacity: 1,
-          transform: 'translateY(0%)'
-        }))
-      ]),
-      transition('* => void', [
-        animate('0.3s', style({
-          opacity: 0,
-          transform: 'translateY(100%)'
-        }))
-      ])
-    ]),
-    trigger('showPageSubTitle', [
-      transition('void => *', [
-        style({
-          opacity: 0,
-          transform: 'translateY(100%)'
-        }),
-        animate('0.4s', style({
-          opacity: 1,
-          transform: 'translateY(0%)'
-        }))
-      ]),
-      transition('* => void', [
-        animate('0.3s', style({
-          opacity: 0,
-          transform: 'translateY(100%)'
-        }))
-      ])
-    ])
+            query('.page-container', [
+              style({
+                opacity: 0,
+                transform: 'translateY(100%)'
+              }), animate('0.3s',
+                style({
+                  opacity: 1,
+                  transform: 'translateY(0%)'
+                }))], {optional: true}),
+            query('.path', [
+              style({
+                opacity: 0,
+                transform: 'translateY(100%)'
+              }), animate('0.3s',
+                style({
+                  opacity: 1,
+                  transform: 'translateY(0%)'
+                }))], {optional: true}),
+          ])),
+        transition(':leave',
+          group([
+            query('.page-container', [
+              style({
+                opacity: 1,
+                transform: 'translateY(0%)'
+              }), animate('0.3s',
+                style({
+                  opacity: 0,
+                  transform: 'translateY(100%)'
+                }))], {optional: true}),
+            query('.path', [
+              style({
+                opacity: 1,
+                transform: 'translateY(0%)'
+              }), animate('0.3s',
+                style({
+                  opacity: 0,
+                  transform: 'translateY(100%)'
+                }))], {optional: true}),
+          ]))
+      ]
+    )
   ]
 })
 export class PageHeaderComponent implements OnInit {
-  @HostBinding('@routerAnimations') showhideme;
+  @HostBinding('@showHideHeader') showHideHeader;
 
   @Input() pagePath: any;
 
