@@ -1,6 +1,6 @@
 import {Component, HostBinding, Input} from '@angular/core';
 import {Router} from '@angular/router';
-import {animate, group, query, state, style, transition, trigger} from '@angular/animations';
+import {animate, state, style, transition, trigger} from '@angular/animations';
 
 @Component({
   selector: 'ebs-header',
@@ -16,31 +16,30 @@ import {animate, group, query, state, style, transition, trigger} from '@angular
         background: '#0049b1',
         color: '#ffffff'
       })),
-      transition('1 => 0', [
-        group([
-          query('.arrowbefore', style({
-            'border-top': '30px solid transparent',
-            'border-bottom': '30px solid transparent'
-          })),
-          query('.arrownext', style({
-            'border-top': '25px solid transparent',
-            'border-bottom': '25px solid transparent'
-          })),
-          animate('300ms')
-        ])
-      ]),
-      transition('0 => 1', [
-        group([
-          query('.arrowbefore', style({
-            'border-top': '30px solid #0049b1',
-            'border-bottom': '30px solid #0049b1'
-          })),
-          query('.arrownext', style({
-            'border-left': '25px solid #0049b1'
-          })),
-          animate('300ms')
-        ])
-      ])
+      transition('1 => 0', animate('300ms')),
+      transition('0 => 1', animate('300ms'))
+    ]),
+    trigger('selecteFirstdNot', [
+      state('0', style({
+        'border-top': '30px solid #eeeeee',
+        'border-bottom': '30px solid #eeeeee'
+      })),
+      state('1', style({
+        'border-top': '30px solid #0049b1',
+        'border-bottom': '30px solid #0049b1'
+      })),
+      transition('1 => 0', animate('300ms')),
+      transition('0 => 1', animate('300ms'))
+    ]),
+    trigger('selecteLastdNot', [
+      state('0', style({
+        'border-left': '25px solid #eeeeee'
+      })),
+      state('1', style({
+        'border-left': '25px solid #0049b1'
+      })),
+      transition('1 => 0', animate('300ms')),
+      transition('0 => 1', animate('300ms'))
     ])
   ]
 })
@@ -49,13 +48,12 @@ export class HeaderComponent {
   @Input() Index: string;
   @Input() Pattern: string[];
 
-  @HostBinding('@selectedNot')
-  get selectedNot(): boolean {
+  @HostBinding('class.selected')
+  get selected(): boolean {
     return this.Pattern?.indexOf(this.router.url) > -1;
   }
 
-  @HostBinding('class.selected')
-  get selected(): boolean {
+  isSelected(): boolean {
     return this.Pattern?.indexOf(this.router.url) > -1;
   }
 
