@@ -1,22 +1,29 @@
 import {Injectable} from '@angular/core';
-import {Router, CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree} from '@angular/router';
+import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router, CanLoad, Route, UrlSegment} from '@angular/router';
 import {AuthService} from '../services/auth.service';
 
-@Injectable()
-export class AuthGuardService implements CanActivate {
-
-  constructor(private router: Router, private authService: AuthService) {
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuardService implements CanActivate, CanLoad {
+  constructor(private router: Router,
+              private authService: AuthService) {
   }
 
-  canActivate(route: ActivatedRouteSnapshot,
-              state: RouterStateSnapshot): boolean | UrlTree {
+  async canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<boolean | UrlTree> {
+    const roles = route.data?.roles as Array<string>;
+    const isInRole = await this.authService.userInRole(roles);
 
-    if (this.authService.isUserLoggedIn()) {
-      const roles = route.data.roles as Array<string>;
-      return (roles !== null && this.authService.userInRole(roles));
+    if (isInRole) {
+      return isInRole;
+    } else {
+      await this.router.navigate(['anmelden'], {queryParams: {retUrl: state.url}});
+      return false;
     }
+  }
 
-    this.router.navigate(['anmelden'], {queryParams: {retUrl: state.url}});
-    return false;
+  async canLoad(route: Route, segments: UrlSegment[]): Promise<boolean> {
+    const roles = route.data?.roles as Array<string>;
+    return await this.authService.userInRole(roles);
   }
 }
